# src/swagger.json
{
	"components": {
...
		"schemas": {
			"AddressRes": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					},
					"default": {
						"type": "boolean"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver",
					"_id",
					"default"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateAddressDTO": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					}
				},
				"required": [
					"province",
					"district",
					"ward",
					"specificAddress",
					"phone",
					"receiver"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_CreateAddressDTO_": {
				"properties": {
					"province": {
						"type": "string"
					},
					"district": {
						"type": "string"
					},
					"ward": {
						"type": "string"
					},
					"specificAddress": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"receiver": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"UpdateAddressDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_CreateAddressDTO_"
					},
					{
						"properties": {
							"default": {
								"type": "boolean"
							}
						},
						"type": "object"
					}
				]
			},
...
			"CommentResDTO": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"user": {
						"properties": {
							"avatar": {
								"type": "string"
							},
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"avatar",
							"name",
							"_id"
						],
						"type": "object"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"content",
					"rate",
					"user",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"IProductModel": {
				"$ref": "#/components/schemas/FlattenMaps_T_"
			},
...
			"ProdutResDTO": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"name",
					"category",
					"price",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductUpdateRes": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"name",
					"category",
					"price",
					"quantity",
					"description",
					"thumbnail",
					"images"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductCard": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"thumbnail": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"name",
					"price",
					"category",
					"thumbnail"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductDetail": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"thumbnail": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"comments": {
						"items": {
							"$ref": "#/components/schemas/CommentResDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"_id",
					"name",
					"price",
					"category",
					"thumbnail",
					"images",
					"quantity",
					"rating",
					"description",
					"comments"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ICreateUser.name_": {
				"properties": {
					"name": {
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserResDTO": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Pick_ICreateUser.name_"
					},
					{
						"properties": {
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"_id"
						],
						"type": "object"
					}
				]
			},
			"ICreateUser": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"username",
					"password",
					"avatar"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ISignJWT": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"userId",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WishlistItem": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"thumbnail": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isStock": {
						"type": "boolean"
					}
				},
				"required": [
					"_id",
					"thumbnail",
					"name",
					"price",
					"isStock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HttpStatus": {
				"enum": [
					200,
					201,
					202,
					400,
					401,
					403,
					404,
					406,
					408,
					500,
					501,
					502,
					503
				],
				"type": "number"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "future-be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/addresses": {
			"post": {
				"operationId": "CreateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AddressRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateAddressDTO"
							}
						}
					}
				}
			}
		},
		"/addresses/{id}": {
			"put": {
				"operationId": "UpdateAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/AddressRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAddressDTO"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Addresses"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/products": {
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"images": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"name",
									"category",
									"quantity",
									"price",
									"description",
									"images"
								]
							}
						}
					}
				}
			}
		},
		"/products/{productId}": {
			"put": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/IProductModel"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"updateImageField": {
										"type": "string"
									},
									"images": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/ProductDetail"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/products/pagination": {
			"get": {
				"operationId": "GetProductsPagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"properties": {
														"numOfProds": {
															"type": "number",
															"format": "double"
														},
														"products": {
															"items": {
																"$ref": "#/components/schemas/ProdutResDTO"
															},
															"type": "array"
														}
													},
													"required": [
														"numOfProds",
														"products"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/products/update/{productId}": {
			"get": {
				"operationId": "GetProductForUpdate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/ProductUpdateRes"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/products/filter": {
			"get": {
				"operationId": "FilterProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/ProductCard"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "category",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "GetTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateUser"
							}
						}
					}
				}
			}
		},
		"/users/sign-jwt-token": {
			"post": {
				"operationId": "SignJwtToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISignJWT"
							}
						}
					}
				}
			}
		},
		"/users/addresses": {
			"get": {
				"operationId": "GetMyAddresses",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/AddressRes"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/wishlist": {
			"get": {
				"operationId": "GetWishlist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"items": {
														"$ref": "#/components/schemas/WishlistItem"
													},
													"type": "array"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "FindUserById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/UserResDTO"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/wishlist/{productId}": {
			"post": {
				"operationId": "InsertWishlistItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/WishlistItem"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteWishlistItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"data": {
													"type": "string"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {
													"$ref": "#/components/schemas/HttpStatus"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"data",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}/cart": {
			"get": {
				"operationId": "GetCart",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}